Move table-modal into dashboard element, fix format and multi year barchart
