+++ frontend/docs/components/table.mdx
@@ -3,7 +3,7 @@ name: Table
 menu: Components
 ---
 import { PropsTable, Playground } from 'docz'
-import Table from 'react-components/table-modal/table';
+import Table from 'react-components/dashboard-element/dashboard-widget/table-modal//table';
 import 'styles/_docz.scss';
 import data from '../mock-data/table-data.json';
