+++ frontend/scripts/react-components/dashboard-element/dashboard-widget/table-modal/table-modal.component.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Heading from 'react-components/shared/heading/heading.component';
 import Button from 'react-components/shared/button';
-import Table from 'react-components/dashboard-element/dashboard-widget/table-modal/table';
+import Table from 'react-components/dashboard-element/dashboard-widget/table-modal//table';
 import 'react-components/dashboard-element/dashboard-widget/table-modal/table-modal.scss';
 
 function TableModal({ title, tableData }) {
