+++ frontend/scripts/react-components/dashboard-element/dashboard-widget/table-modal/table-modal.scss
@@ -1,0 +1,18 @@
+@import 'styles/settings';
+
+.c-table-modal {
+  width: 820px;
+  padding: 30px 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-end;
+
+  .table {
+    margin-top: 40px;
+  }
+
+  .table-modal-footer {
+    padding: 25px 0 0;
+  }
+}
