+++ frontend/scripts/react-components/dashboard-element/dashboard-widget/table-modal/table/table.component.jsx
